+++ pkgs/coverage/bin/test_with_coverage.dart
@@ -6,8 +6,10 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:args/args.dart';
+import 'package:coverage/src/coverage_options.dart';
 import 'package:coverage/src/util.dart'
     show StandardOutExtension, extractVMServiceUri;
+import 'package:meta/meta.dart';
 import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as path;
 
@@ -50,37 +52,45 @@ void _watchExitSignal(ProcessSignal signal) {
   });
 }
 
-ArgParser _createArgParser() => ArgParser()
+ArgParser _createArgParser(CoverageOptions defaultOptions) => ArgParser()
   ..addOption(
     'package',
     help: 'Root directory of the package to test.',
-    defaultsTo: '.',
+    defaultsTo: defaultOptions.packageDirectory,
   )
   ..addOption(
     'package-name',
     help: 'Name of the package to test. '
         'Deduced from --package if not provided.',
+    defaultsTo: defaultOptions.packageName,
   )
   ..addOption('port', help: 'VM service port.', defaultsTo: '8181')
   ..addOption(
     'out',
+    defaultsTo: defaultOptions.outputDirectory,
     abbr: 'o',
     help: 'Output directory. Defaults to <package-dir>/coverage.',
   )
-  ..addOption('test', help: 'Test script to run.', defaultsTo: 'test')
+  ..addOption('test',
+      help: 'Test script to run.', defaultsTo: defaultOptions.testScript)
   ..addFlag(
     'function-coverage',
     abbr: 'f',
-    defaultsTo: false,
+    defaultsTo: defaultOptions.functionCoverage,
     help: 'Collect function coverage info.',
   )
   ..addFlag(
     'branch-coverage',
     abbr: 'b',
-    defaultsTo: false,
+    defaultsTo: defaultOptions.branchCoverage,
     help: 'Collect branch coverage info.',
   )
+  ..addOption(
+    'fail-under',
+    help: 'Fail if coverage is less than the given percentage (0-100)',
+  )
   ..addMultiOption('scope-output',
+      defaultsTo: defaultOptions.scopeOutput,
       help: 'restrict coverage results so that only scripts that start with '
           'the provided package path are considered. Defaults to the name of '
           'the package under test.')
@@ -95,7 +105,8 @@ class Flags {
     this.testScript,
     this.functionCoverage,
     this.branchCoverage,
-    this.scopeOutput, {
+    this.scopeOutput,
+    this.failUnder, {
     required this.rest,
   });
 
@@ -107,11 +118,14 @@ class Flags {
   final bool functionCoverage;
   final bool branchCoverage;
   final List<String> scopeOutput;
+  final String? failUnder;
   final List<String> rest;
 }
 
-Future<Flags> _parseArgs(List<String> arguments) async {
-  final parser = _createArgParser();
+@visibleForTesting
+Future<Flags> parseArgs(
+    List<String> arguments, CoverageOptions defaultOptions) async {
+  final parser = _createArgParser(defaultOptions);
   final args = parser.parse(arguments);
 
   void printUsage() {
@@ -138,7 +152,7 @@ ${parser.usage}
     exit(0);
   }
 
-  final packageDir = path.canonicalize(args['package'] as String);
+  final packageDir = path.normalize(path.absolute(args['package'] as String));
   if (!FileSystemEntity.isDirectorySync(packageDir)) {
     fail('--package is not a valid directory.');
   }
@@ -161,12 +175,14 @@ ${parser.usage}
     args['function-coverage'] as bool,
     args['branch-coverage'] as bool,
     args['scope-output'] as List<String>,
+    args['fail-under'] as String?,
     rest: args.rest,
   );
 }
 
 Future<void> main(List<String> arguments) async {
-  final flags = await _parseArgs(arguments);
+  final defaultOptions = CoverageOptionsProvider().coverageOptions;
+  final flags = await parseArgs(arguments, defaultOptions);
   final outJson = path.join(flags.outDir, 'coverage.json');
   final outLcov = path.join(flags.outDir, 'lcov.info');
 
@@ -224,6 +240,7 @@ Future<void> main(List<String> arguments) async {
     outJson,
     '-o',
     outLcov,
+    if (flags.failUnder != null) '--fail-under=${flags.failUnder}',
   ]);
   exit(0);
 }