+++ pkgs/code_builder/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.10.2-wip
+
+* Upgrade `dart_style` and `source_gen` to remove `package:macros` dependency.
+* Require Dart `^3.6.0` due to the upgrades.
+* Support `Expression.newInstanceNamed` with empty name
+
 ## 4.10.1
 
 * Require Dart `^3.5.0`