+++ pkgs/html/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.15.6
+
+- Performance improvements.
+- No longer generate any error spans if generateSpans is false.
+- Fixed a TypeError in nth-child with non numeric value (e.g. `nth-child(even)`)
+
+## 0.15.5+1
+
+- Support "ambiguous ampersand" in attribute values.
+
 ## 0.15.5
 
 - Require Dart `3.2`.